requirements.txt
DateTime==4.3
prettytable==2.2.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyTelegramBotAPI==4.1.0
python-telegram-bot-pagination==0.0.2
pytz==2021.3
requests==2.26.0
urllib3==1.26.7
wcwidth==0.2.5
zope.interface==5.4.0
